requirements.txt
streamlit
qdrant-client
sentence-transformers
PyPDF2
python-docx
pandas
fpdf
google-generativeai
langchain==0.1.14
langchain-text-splitters==0.0.1
